cli: separate cli binaries from libraries to make code more testable
