+++ cli/connect/run-local.sh
