+++ cli/mcp/.vscode/settings.json
