+++ cli/mcp/Dockerfile
