+++ cli/mcp/Makefile
