+++ cli/mcp/logging_middleware.go
