+++ cli/mcp/main.go
