+++ cli/mcp/run-local.sh
