+++ cli/mcp/server.go
