+++ cli/mcp/server_test.go
