+++ cli/mcp/tools.go
