+++ cli/proxy/run-local.sh
