+++ cli/proxy/run-main.sh
