+++ cli/taskworker/run-main.sh
